refactor conversations to tab style and list messages and create messages based on conversation
